setup shared dependencies
