+++ products/webpack.config.js
@@ -15,6 +15,7 @@ module.exports = {
       exposes: {
         "./ProductsIndex": "./src/index.js",
       },
+      shared: ["faker"],
     }),
   ],
 };
